+++ debug-tools-base/src/main/java/io/github/future0923/debug/tools/base/hutool/core/annotation/package-info.java
@@ -1,0 +1,7 @@
+/**
+ * 注解包，提供增强型注解和注解工具类
+ *
+ * @author looly
+ *
+ */
+package io.github.future0923.debug.tools.base.hutool.core.annotation;