+++ debug-tools-base/src/main/java/io/github/future0923/debug/tools/base/hutool/core/clone/package-info.java
@@ -1,0 +1,7 @@
+/**
+ * 克隆封装
+ *
+ * @author looly
+ *
+ */
+package io.github.future0923.debug.tools.base.hutool.core.clone;