+++ debug-tools-base/src/main/java/io/github/future0923/debug/tools/base/hutool/core/convert/impl/CurrencyConverter.java
@@ -1,0 +1,21 @@
+package io.github.future0923.debug.tools.base.hutool.core.convert.impl;
+
+import io.github.future0923.debug.tools.base.hutool.core.convert.AbstractConverter;
+
+import java.util.Currency;
+
+/**
+ * 货币{@link Currency} 转换器
+ *
+ * @author Looly
+ * @since 3.0.8
+ */
+public class CurrencyConverter extends AbstractConverter<Currency> {
+	private static final long serialVersionUID = 1L;
+
+	@Override
+	protected Currency convertInternal(Object value) {
+		return Currency.getInstance(convertToStr(value));
+	}
+
+}