+++ debug-tools-base/src/main/java/io/github/future0923/debug/tools/base/hutool/core/exceptions/DependencyException.java
@@ -1,0 +1,38 @@
+package io.github.future0923.debug.tools.base.hutool.core.exceptions;
+
+import io.github.future0923.debug.tools.base.hutool.core.exceptions.ExceptionUtil;
+import io.github.future0923.debug.tools.base.hutool.core.util.StrUtil;
+
+/**
+ * 依赖异常
+ *
+ * @author xiaoleilu
+ * @since 4.0.10
+ */
+public class DependencyException extends RuntimeException {
+	private static final long serialVersionUID = 8247610319171014183L;
+
+	public DependencyException(Throwable e) {
+		super(ExceptionUtil.getMessage(e), e);
+	}
+
+	public DependencyException(String message) {
+		super(message);
+	}
+
+	public DependencyException(String messageTemplate, Object... params) {
+		super(StrUtil.format(messageTemplate, params));
+	}
+
+	public DependencyException(String message, Throwable throwable) {
+		super(message, throwable);
+	}
+
+	public DependencyException(String message, Throwable throwable, boolean enableSuppression, boolean writableStackTrace) {
+		super(message, throwable, enableSuppression, writableStackTrace);
+	}
+
+	public DependencyException(Throwable throwable, String messageTemplate, Object... params) {
+		super(StrUtil.format(messageTemplate, params), throwable);
+	}
+}