+++ debug-tools-base/src/main/java/io/github/future0923/debug/tools/base/hutool/core/io/checksum/crc16/package-info.java
@@ -1,0 +1,7 @@
+/**
+ * CRC16相关算法封装为Checksum
+ *
+ * @author looly
+ *
+ */
+package io.github.future0923.debug.tools.base.hutool.core.io.checksum.crc16;