+++ debug-tools-base/src/main/java/io/github/future0923/debug/tools/base/hutool/core/io/copy/package-info.java
@@ -1,0 +1,7 @@
+/**
+ * IO流拷贝相关封装相关封装
+ *
+ * @author looly
+ * @since 5.7.8
+ */
+package io.github.future0923.debug.tools.base.hutool.core.io.copy;