+++ debug-tools-base/src/main/java/io/github/future0923/debug/tools/base/hutool/core/lang/generator/UUIDGenerator.java
@@ -1,0 +1,16 @@
+package io.github.future0923.debug.tools.base.hutool.core.lang.generator;
+
+import io.github.future0923.debug.tools.base.hutool.core.util.IdUtil;
+
+/**
+ * UUID生成器
+ *
+ * @author looly
+ * @since 5.4.3
+ */
+public class UUIDGenerator implements Generator<String> {
+	@Override
+	public String next() {
+		return IdUtil.fastUUID();
+	}
+}