+++ debug-tools-base/src/main/java/io/github/future0923/debug/tools/base/hutool/json/serialize/JSONObjectSerializer.java
@@ -1,0 +1,12 @@
+package io.github.future0923.debug.tools.base.hutool.json.serialize;
+
+import io.github.future0923.debug.tools.base.hutool.json.JSONObject;
+
+/**
+ * 对象的序列化接口，用于将特定对象序列化为{@link JSONObject}
+ * @param <V> 对象类型
+ * 
+ * @author Looly
+ */
+@FunctionalInterface
+public interface JSONObjectSerializer<V> extends JSONSerializer<JSONObject, V>{}